refactored imports to support ts-node with ESM (TESTING REQUIRED)
