+++ packages/pure/package.json
@@ -7,6 +7,7 @@
   },
   "license": "ISC",
   "main": "dist/index.js",
+  "type": "module",
   "types": "dist/index.d.ts",
   "scripts": {
     "clean": "rm -rf ./dist",
