+++ packages/pure/src/device/constants/index.ts
@@ -3,6 +3,6 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./device-type.enum"
-export * from "./manufactures.enum"
-export * from "./product-identifier"
+export * from "./device-type.enum.js"
+export * from "./manufactures.enum.js"
+export * from "./product-identifier.js"
