+++ packages/pure/src/device/descriptors/device-descriptor.ts
@@ -3,8 +3,8 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { Manufacture, VendorID, ProductID, DeviceType } from "../constants"
-import { PureStrategy, HarmonyStrategy } from "../strategies"
+import { Manufacture, VendorID, ProductID, DeviceType } from "../constants/index.js"
+import { PureStrategy, HarmonyStrategy } from "../strategies/index.js"
 
 export interface DeviceDescriptor {
   manufacturer: Manufacture
