+++ packages/pure/src/device/descriptors/index.ts
@@ -3,6 +3,6 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./mudita-pure.descriptor"
-export * from "./mudita-harmony.descriptor"
-export * from "./device-descriptor"
+export * from "./mudita-pure.descriptor.js"
+export * from "./mudita-harmony.descriptor.js"
+export * from "./device-descriptor.js"
