+++ packages/pure/src/device/index.ts
@@ -3,11 +3,11 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./base-device"
-export * from "./device"
-export * from "./device.types"
-export * from "./constants"
-export * from "./descriptors"
-export * from "./services"
-export * from "./validators"
-export * from "./strategies"
+export * from "./base-device.js"
+export * from "./device.js"
+export * from "./device.types.js"
+export * from "./constants/index.js"
+export * from "./descriptors/index.js"
+export * from "./services/index.js"
+export * from "./validators/index.js"
+export * from "./strategies/index.js"
