+++ packages/pure/src/device/validators/index.ts
@@ -3,4 +3,4 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./port-info.validator"
+export * from "./port-info.validator.js"
