+++ packages/pure/src/device/validators/port-info.validator.ts
@@ -4,7 +4,7 @@
  */
 
 import { PortInfo } from "serialport"
-import { MuditaPureDescriptor, MuditaHarmonyDescriptor } from "../descriptors"
+import { MuditaPureDescriptor, MuditaHarmonyDescriptor } from "../descriptors/index.js"
 
 export class PortInfoValidator {
   static eligibleDevices = [MuditaPureDescriptor, MuditaHarmonyDescriptor]
