+++ packages/pure/src/endpoints/index.ts
@@ -3,11 +3,11 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./backup.types"
-export * from "./contact.types"
-export * from "./device-info.types"
-export * from "./device-update.types"
-export * from "./file-system"
-export * from "./messages.types"
-export * from "./outbox.types"
-export * from "./restore.types"
+export * from "./backup.types.js"
+export * from "./contact.types.js"
+export * from "./device-info.types.js"
+export * from "./device-update.types.js"
+export * from "./file-system.js"
+export * from "./messages.types.js"
+export * from "./outbox.types.js"
+export * from "./restore.types.js"
