+++ packages/pure/src/endpoints/outbox.types.ts
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { Endpoint, Method, RequestConfig } from "../device"
+import { Endpoint, Method, RequestConfig } from "../device/index.js"
 
 export enum OutboxEntryType {
   Message = 1,
