+++ packages/pure/src/endpoints/restore.types.ts
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { Endpoint, Method, RequestConfig } from "../device"
+import { Endpoint, Method, RequestConfig } from "../device/index.js"
 
 type backupId = string
