+++ packages/pure/src/formatter/formatter-factory.ts
@@ -3,8 +3,8 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { PureV1Formatter } from "./pure-v1.formatter"
-import { Formatter } from "./formatter"
+import { PureV1Formatter } from "./pure-v1.formatter.js"
+import { Formatter } from "./formatter.js"
 
 const formatterMap: { [key: number]: Formatter } = {
   1: new PureV1Formatter(),
