+++ packages/pure/src/index.ts
@@ -3,12 +3,12 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./device"
-export * from "./logger"
-export * from "./endpoints"
-export * from "./formatter"
-export * from "./device-manager"
-export * from "./timeout"
+export * from "./device/index.js"
+export * from "./logger/index.js"
+export * from "./endpoints/index.js"
+export * from "./formatter/index.js"
+export * from "./device-manager.js"
+export * from "./timeout.js"
 
-import PureNode from "./device-manager"
+import PureNode from "./device-manager.js"
 export default PureNode
