+++ packages/pure/src/logger/logger-factory.ts
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import Logger, { PureLogger } from "./logger"
+import Logger, { PureLogger } from "./logger.js"
 
 let logger: PureLogger
