+++ packages/pure/src/logger/logger.test.ts
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import Logger, { PureLogger, ConsoleLogger } from "./logger"
+import Logger, { PureLogger, ConsoleLogger } from "./logger.js"
 
 let logger: PureLogger
